visualizer/requirements.txt
streamlit
streamlit-autorefresh
plotly
networkx
matplotlib
google-cloud-pubsub
protobuf 